final merge

+++ controllers/transaction.controller.js
@@ -1,15 +1,31 @@
 const Transaction = require('../schemas/transaction.schema');
-
+const mongoose = require("mongoose");
+const {getUserBalance, updateUserBalance} = require("./user.controller");
 const addTransaction = async (type, category, amount, date, comment, owner) => {
     return Transaction.create({type, category, amount, date, comment, owner});
 }
 
 const deleteTransaction = async (transactionId) => {
-    return Transaction.findOneAndDelete({_id: transactionId})
+    const transaction = await getTransactionById(transactionId);
+    if (!transaction) {
+        throw new Error('Transaction does not exist');
+    }
+    await Transaction.findOneAndDelete({ _id: transactionId });
+    const type = transaction.type;
+    const amount = transaction.amount;
+    const owner = transaction.owner;
+    let balance = await getUserBalance(owner);
+    if (type === 'income') {
+        balance -= amount;
+    }
+    if (type === 'expense') {
+        balance += amount;
+    }
+    await updateUserBalance(owner, balance);
 }
 
 const updateTransaction = async (transactionId, type, category, amount, date, comment, owner) => {
-    return Transaction.updateOne({_id: transactionId},{type, category, amount, date, comment, owner});
+    return Transaction.findOneAndReplace({_id: transactionId}, {type, category, amount, date, comment, owner});
 }
 
 const getTransactionById = async (transactionId) => {
@@ -17,7 +33,8 @@ const getTransactionById = async (transactionId) => {
 }
 
 const getUserTransactions = async (userId) => {
-    return Transaction.find({userId});
+    const userIdToObjectId = new mongoose.Types.ObjectId(userId);
+    return Transaction.find({owner: userIdToObjectId});
 }
 
 const getUserStatisticsByDate = async (userId, transactionsDate) => {
@@ -42,6 +59,20 @@ const getUserStatisticsByDate = async (userId, transactionsDate) => {
     return [transactionsFilteredByDate, sumOfIncome, sumOfExpense];
 }
 
+const getTransactionsAmountsDifference = async (transactionId, type, amount, owner) => {
+    const primaryTransaction = await getTransactionById(transactionId);
+    let balance = await getUserBalance(owner);
+    let difference = 0;
+    if (type === 'income') {
+        difference = amount - primaryTransaction.amount;
+    }
+    if (type === 'expense') {
+        difference = primaryTransaction.amount - amount;
+    }
+    balance += difference;
+    return balance;
+}
+
 module.exports = {
     addTransaction,
     getTransactionById,
@@ -49,4 +80,5 @@ module.exports = {
     getUserStatisticsByDate,
     updateTransaction,
     deleteTransaction,
+    getTransactionsAmountsDifference,
 }
